refactor: rename stream buffer module
